pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "ai-palindromikisa"
version = "0.1.0"
description = "AI Palindrome Benchmark - evaluating LLM capability for creating palindromes"
requires-python = ">=3.10"
dependencies = [
    "click",
    "click-default-group>=1.2.3",
    "livereload",
    "llm",
    "llm-anthropic",
    "llm-gemini",
    "llm-openrouter",
    "plotext",
    "pyyaml",
    "requests",
    "rich",
    "ruamel-yaml>=0.18.16",
    "tenacity",
]

[dependency-groups]
dev = [
    "mypy",
    "pytest",
    "pytest-mock",
]

[project.scripts]
ai-palindromikisa = "ai_palindromikisa.cli:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short -m 'not live'"
markers = [
    "live: tests that call real LLM APIs (deselected by default, run with: pytest -m live)",
]

[tool.pyright]
reportAny = "none"
reportMissingParameterType = "none"
reportUnknownArgumentType = "none"
reportUnknownLambdaType = "none"
reportUnknownMemberType = "none"
reportUnknownParameterType = "none"
reportUnknownVariableType = "none"
reportUnusedCallResult = "none"
